backend/requirements.txt
aiofiles==0.5.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
alembic==1.4.2
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.1.7
brotlipy==0.7.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
databases==0.4.3
dnspython==1.16.0
ecdsa==0.15
email-validator==1.1.1
Fabric3==1.13.1.post1
fastapi==0.59.0
fastapi-utils==0.2.1
h11==0.9.0
idna==2.9
invoke==1.5.0
itsdangerous==2.0.1
jieba==0.42.1
kiwisolver==1.3.1
loguru==0.5.1
lxml==4.6.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.4.2
motor==2.4.0
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
paramiko==2.7.2
passlib==1.7.2
pathlib2==2.3.5
Pillow==8.2.0
psutil==5.8.0
pyasn1==0.4.8
pycosat==0.6.3
pycparser
pydantic==1.5.1
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==0.9.3
PyNaCl==1.4.0
pyOpenSSL
pyparsing==2.4.7
PySocks
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-multipart==0.0.5
pytz==2021.1
requests
rsa==4.6
ruamel-yaml==0.15.87
SQLAlchemy==1.3.17
starlette==0.13.4
typing-extensions==3.10.0.0
uvicorn==0.11.5
websockets==8.1
yarl==1.6.3
aioredis
fastapi-cache2
pycryptodome
redis